Safari-609.1.11.1/LayoutTests/editing/selection/ios/hide-selection-in-tiny-contenteditable-expected.txt

Focus the editor
abcdefg
Verifies that selection UI is suppressed when the editable root is extremely small. To manually test, tap on the button above and verify that (1) the editable element is focused, and (2) selection handles are not shown.

On success, you will see a series of "PASS" messages, followed by "TEST COMPLETE".

After focus, the caret rect is empty
After zooming in, the caret rect is empty
After making editor large, the caret rect is empty
After making editor opaque, the caret rect is (left=21, top=100, width=2, height=19)
After making editor tiny again, the caret rect is empty
After making editor transparent again, the caret rect is empty
After making editor large again, the caret rect is empty
After making editor opaque again, the caret rect is (left=50, top=100, width=2, height=19)
PASS successfullyParsed is true

TEST COMPLETE
